Removed default instantiation of local storage and added it as a module instead.
