+++ lib/Storage.js
@@ -50,4 +50,6 @@ function Storage() {
   this.writable = true;
   this.configurable = true;
   this.enumerable = true;
+
+  return oStorage;
 }
