+++ lib/Window.js
@@ -1,7 +1,6 @@
 "use strict";
 
 const Element = require("./Element");
-const Storage = require("./Storage");
 const getLocation = require("./getLocation");
 const Navigator = require("./Navigator");
 const url = require("url");
@@ -54,8 +53,6 @@ module.exports = function Window(resp, windowObjects, innerWidth = 760, innerHei
   Object.defineProperty(window, "self", {
     get: () => window
   });
-
-  Object.defineProperty(window, "localStorage", new Storage());
 
   return window;
