+++ lib/index.js
@@ -6,6 +6,7 @@ const Fetch = require("./Fetch");
 const IntersectionObserver = require("./IntersectionObserver");
 const MutationObserver = require("./MutationObserver");
 const Window = require("./Window");
+const Storage = require("./Storage");
 
 module.exports = {
   Compiler,
@@ -14,4 +15,5 @@ module.exports = {
   IntersectionObserver,
   MutationObserver,
   Window,
+  Storage
 };
